requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
neo4j==5.15.0
python-dotenv==1.0.0
pydantic==2.5.0

# Frontend dependencies
streamlit==1.28.1
requests==2.31.0

# Chatbot dependencies
openai==1.3.0
